Address Python usage graph PR comments
